backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
exceptiongroup==1.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.47.0
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
pillow==11.3.0
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.13.1
pydantic_core==2.46.1
pyparsing==3.3.2
python-dotenv==1.2.1
requests==2.32.5
tenacity==9.1.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.9
httpx==0.28.1
